+++ CHANGELOG.md
@@ -6,6 +6,15 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - Feature: Add `IconList`.
 
+## [1.21.0] - 2019-03-11
+
+Features:
+- Add `Video` component.
+- Add `getReactElementsWithoutType` helper.
+
+Fix:
+- Add missing `href` props to `SimpleCard`.
+
 ## [1.20.0] - 2019-03-08
 
 Feature: